requirements.txt
streamlit
pandas
plotly
python-docx
openpyxl
kaleido
loguru